Issue 675: heap-buffer-overflow Error In tinyxml2.cpp
When XMLDocument::Parse(const char* xml, size_t nBytes) is called with
the 'nBytes' parameter set to (size_t)-1, the function uses strlen to
get the size of the 'xml' parameter.  Incidentally, that xml parameter
must end with the '\0' character.

When that function is called with an 'xml' string that does *NOT* end
with a '\0' character and if 'nBytes' was set to (size_t)-1, then the
strlen() function called on the 'xml' parameter would overflow in an
unpredictable way.

This patch adds an overload member function
XMLDocument::Parse(const char* xml) which requires that the xml buffer
has a maximal size of 4K bytes, or ends-up with a '\0'.  In any case,
the xml buffer won't be able to be greater in size than 4K bytes.

For an xml buffer of an arbitrary size, the user would have to use the
XMLDocument::Parse(const char* xml, size_t nBytes) function.

The patch also adjusts the xmltest.cpp test to make it specify the xml
buffer size for the test that requires it.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>